+++ Samples~/Scripts/SaintsDecimalExample/SaintsDecimalField.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsDecimalExample
+{
+    public class SaintsDecimalField : MonoBehaviour
+    {
+        public SaintsDecimal saintsDecimal;
+
+        private void Awake()
+        {
+            // implicit converting supported
+            decimal convertToDecimal = saintsDecimal;
+            decimal add = decimal.One + saintsDecimal;
+        }
+    }
+}